Root_Agent_2/requirements.txt
google-adk
google-generativeai
pydantic
python-dotenv